+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## [Unreleased]
+
+### Added
+
+- `retry` in common utils
+- `LIBRARY_ERROR_SCOPE` constant
+
 ## [2.4.0]
 
 ### Added